+++ src/lib.rs
@@ -28,7 +28,8 @@
 //! let frame = 0.0; // Arbitrary number chosen. Ensure it's a valid frame!
 //! let transform = vello::kurbo::Affine::IDENTITY;
 //! let alpha = 1.0;
-//! let scene = renderer.render_to_vello_scene(&composition, frame, transform, alpha);
+//! let images = std::collections::HashMap::new();
+//! let scene = renderer.render_to_vello_scene(&composition, &images, frame, transform, alpha);
 //! ```
 //!
 //! # Unsupported features